+++ teneto/networkmeasures/shortest_temporal_path.py
@@ -1,116 +1,176 @@
-"""
-
-Networkmeasures: shortest temporal paths
-
-"""
 
 import numpy as np
-from ..utils import process_input
+from teneto.utils import process_input
+from numba import jit
+import itertools 
+import pandas as pd 
 
+def seqpath_to_path(pairseq, source):
+    # seq must be a path sequence (i.e. possible paths per timepoint)
+    # convert the sequence of pairs to a n x 2 array
+    pairrows = np.reshape(pairseq,[int(len(pairseq)/2), 2])
+    queue = [(0, [0])]
+    # if source is in the first tuple, return
+    if source in pairrows[0]:
+        yield [pairrows[0].tolist()]
+    while queue:
+        # Set the queue
+        (node, path) = queue.pop(0)
+        # Get all remaining possible paths in sequence
+        iterset = set(np.where((pairrows==pairrows[node,0]) | (pairrows==pairrows[node,1]))[0]) - set(range(node+1))
+        for next in iterset:
+            if source in pairrows[next]:
+                yield list(reversed(pairrows[path + [next]].tolist()))
+            else:
+                queue.append((next, path + [next]))
 
-def shortest_temporal_path(tnet, quiet=1):
-    """
-    Calculates the shortest temporal path when all possible routes can be travelled
-    at each time point.
-    Currently only works for binary undirected edges (but can be expanded).
+def shortest_path_from_pairseq(pairseq, source):
+    try:
+        return next(seqpath_to_path(pairseq, source))
+    except StopIteration:
+        return None
+
+def shortest_temporal_path(tnet, steps_per_t='all', i=None, j=None, it=None, minimise='time'):
+    """ 
+    Shortest temporal path
 
     Parameters
-    ----------
+    --------------
 
-    tnet : array, dict 
-        temporal network input (graphlet or contact). Nettype: 'bu'
-    quiet : int (default = 1). 
-        Turn to 0 if you want progress update.
+    tnet : tnet obj, array or dict
+        input network. nettype: bu, bd. 
 
-    Returns
-    --------
-    paths : dict 
-        Shortest temporal paths. Dictionary is in the form, path['paths'][i,j,t] - shortest path for i to reach j starting at time t.
+    steps_per_t : int or str
+        If str, should be 'all'. 
+        How many edges can be travelled during a single time-point. 
 
-    Examples
-    ---------
+    i : list 
+        List of node indicies to restrict analysis. These are nodes the paths start from. Default is all nodes.
 
-    This example shows how to calculate the shortest temporal paths for a network 
+    j : list
+        List of node indicies to restrict analysis. There are nodes the paths end on.  Default is all nodes.
 
-    >>> import teneto 
-    >>> import numpy as np 
+    it : list
+        List of starting time-point indicies to restrict anlaysis. Default is all timepoints.
 
-    Make a network with 3 nodes and 4 time-points. 
+    Returns 
 
-    >>> G = np.zeros([3,3,4])
-    >>> G[0,1,0] = 1
-    >>> G[0,1,2] = 1
-    >>> G[1,2,1] = 1
-    >>> G[1,2,3] = 1
-    >>> G = G + G.transpose([1,0,2])
-
-    Calculting the intercontact times produces a dictionary. 
-
-    >>> paths = teneto.networkmeasures.shortest_temporal_path(G)
-
-    The key 'paths' in the dictionary that is NxNxT. The values 
-    say how long it takes for a path to go from i to j. Values are nans 
-    along the diagonal and when there are no paths available. 
-    If we consider node 0 and 2 at time-point 0, it should take 2 time-points to get there (from node-0 to node-1 at time-point 0 and from node-1 to node-2 at time-point 1).
-
-    >>> paths['paths'][0,2,0]
-    2.0
-
-    While the path from node 2 to 0 at time-point 0 will take 3 time-points (from node-2 to node-1 at time-point 1 and from node-1 to node-0 at time-point 3).
-
-    >>> paths['paths'][2,0,0]
-    3.0
-
-    As paths some node-time-points do not have any paths, you can check when nodes are missing the shortest paths. 
-
-    >>> paths['percentReached'] 
-    0.7916666666666666
-
+    paths : pandas df 
+        Dataframe consisting of information about all the paths found. 
     """
 
-    # Get input type (C or G)
-    # Process input
-    tnet, netinfo = process_input(tnet, ['C', 'G', 'TN'])
+    tnet = process_input(tnet, ['C', 'G', 'TN'], 'TN')
 
-    if netinfo['nettype'] != 'bu':
-        errormsg = ('It looks like your graph is not binary and undirected. '
-                    'Shortest temporal paths can only be calculated for '
-                    'binary undirected networks in Teneto at the moment. '
-                    'If another type is required, please create an issue at: '
-                    'github.com/wiheto/teneto.')
-        raise ValueError(errormsg)
+    # If i, j or it are inputs, process them
+    if i is None: 
+        source_nodes = np.arange(tnet.netshape[0])
+    elif isinstance(i,int): 
+        source_nodes = [i]
+    elif isinstance(i,list): 
+        source_nodes = i    
+    else: 
+        raise ValueError('Unknown i input. Should be None, int or list')
+    if j is None: 
+        target_nodes = np.arange(tnet.netshape[0])
+    elif isinstance(j, int): 
+        target_nodes = [j]
+    elif isinstance(j,list): 
+        target_nodes = j    
+    else: 
+        raise ValueError('Unknown j input. Should be None, int or list')
+    if it is None: 
+        time_points = np.arange(tnet.netshape[1])
+    elif isinstance(it, int): 
+        time_points = [it]
+    elif isinstance(it, list): 
+        time_points = it    
+    else: 
+        raise ValueError('Unknown t input. Should be None, int or list')
 
-    # Preallocate output
-    paths = np.zeros(netinfo['netshape']) * np.nan
-    # Go backwards in time and see if something is reached
-    paths_last_contact = np.zeros(
-        [netinfo['netshape'][0], netinfo['netshape'][1]]) * np.nan
-    for t_ind in list(reversed(range(0, netinfo['netshape'][2]))):
-        if quiet == 0:
-            print('--- Running for time: ' + str(t_ind) + ' ---')
-        fid = np.where(tnet[:, :, t_ind] >= 1)
-        # Update time step
-        # Note to self: Add a conditional to prevent nan warning
-        # that can pop out if no path is there straight away.
-        paths_last_contact += 1
-        # Reset connections present to 1s
-        paths_last_contact[fid[0], fid[1]] = 1
-        # Update nodes with no connections
-        # Nodes to update are nodes with an edge present at the time point
-        for i in np.unique(fid[0]):
-            connections = np.where(paths_last_contact[i, :] == 1)[0]
-            #paths_last_contact_preupdate = np.array(paths_last_contact[i,:])
-            paths_last_contact[i, :] = np.nanmin(
-                paths_last_contact[np.hstack([connections, i]), :], axis=0)
-            # make self connection nan regardless
-            paths_last_contact[i, i] = np.nan
-        paths[:, :, t_ind] = paths_last_contact
-    # Return output
-    paths_dict = {}
-    percentreach = (paths.size - np.sum(np.isnan(paths))) / \
-        (paths.size - (paths.shape[0] * paths.shape[2]))
-    paths_dict['percentReached'] = percentreach
-    paths_dict['paths'] = paths
-    paths_dict['nettype'] = netinfo['nettype']
+    # Two step process. 
+    # First, get what the network can reach per timepoint. 
+    # Second, check all possible sequences of what the network can reach for the shortest sequence.  
+    paths = []
+    for source in source_nodes:
+        for target in target_nodes:
+            if target == source: 
+                pass
+            else:
+                for tstart in time_points:
+                    # Part 1 starts here
+                    ij = [source] 
+                    t = tstart 
+                    step = 1
+                    lenij = 1
+                    pairs = []
+                    stop = 0
+                    while stop == 0:
+                        # Only select i if directed, ij if undirected. 
+                        if tnet.nettype[1] == 'u': 
+                            network = tnet.get_network_when(ij=list(ij),t=t)
+                        elif tnet.nettype[1] == 'd': 
+                            network = tnet.get_network_when(i=list(ij),t=t)
+                        new_nodes = network[['i','j']].values
+                        if len(new_nodes) != 0:
+                            pairs.append(new_nodes.tolist())                 
+                        new_nodes = new_nodes.flatten() 
+                        ij = np.hstack([ij, new_nodes])
+                        ij = np.unique(ij)
+                        if minimise == 'time' and target in ij: 
+                            stop = 1
+                        elif minimise == 'topology' and t == tnet.netshape[1]  and target in ij:
+                            stop = 1
+                        elif t == tnet.netshape[1]:
+                            t = np.nan
+                            ij = [target]
+                            stop = 1
+                        else:
+                            if len(ij) == lenij: 
+                                t += 1
+                                step = 1
+                            elif steps_per_t == 'all':
+                                pass
+                            elif step < steps_per_t: 
+                                step += 1
+                            else: 
+                                t += 1
+                                step = 1
+                        lenij = len(ij)
+                    # correct t for return
+                    t += 1
+                    # Path 2 starts here
+                    path = np.nan
+                    pl = np.nan
+                    for n in itertools.product(*reversed(pairs)): 
+                        a = np.array(n).flatten()
+                        if source not in a or target not in a: 
+                            pass
+                        else:
+                            pathtmp = shortest_path_from_pairseq(a, source)
+                            if pathtmp:
+                                if not isinstance(path, list): 
+                                    path = pathtmp
+                                    pl = len(path)
+                                elif len(pathtmp) < pl:  
+                                    path = pathtmp
+                                    pl = len(path)
+                                elif len(pathtmp) == pl: 
+                                    if isinstance(path[0][0], list):
+                                        if pathtmp in path: 
+                                            pass
+                                        else:
+                                            path.append(pathtmp)
+                                    else: 
+                                        if path == pathtmp:
+                                            pass
+                                        else:
+                                            path = [path, pathtmp]
+                        #elif sourcei < 2 and target in a[:2]:
+                        #    pl = 2
 
-    return paths_dict
+                    paths.append([source,target,tstart,t-tstart,pl,path])
+
+    paths = pd.DataFrame(data=paths, columns=['from', 'to', 't_start', 'temporal-distance', 'topological-distance', 'path includes'])
+    return paths
+
+