+++ teneto/networkmeasures/temporal_closeness_centrality.py
@@ -6,18 +6,22 @@ import numpy as np
 from .shortest_temporal_path import shortest_temporal_path
 
 
-def temporal_closeness_centrality(data):
+def temporal_closeness_centrality(tnet=None,paths=None):
     '''
-    returns temporal closeness centrality per node.
-    As temporalPaths only works with binary undirected edges at the moment,
-     this is required for temporal closeness centrality.
+    Returns temporal closeness centrality per node.
 
     Parameters 
     -----------
 
+    Input should be *either* tnet or paths. 
+
     data : array or dict 
     
-        Temporal network input (graphlet or contact). nettype: 'bu'. Can also be a dictionary of paths (output of TenetoBIDS.networkmeasure.shortest_temporal_paths)
+        Temporal network input (graphlet or contact). nettype: 'bu', 'bd'. 
+        
+    paths : pandas dataframe
+    
+        Output of TenetoBIDS.networkmeasure.shortest_temporal_paths
 
 
     Returns 
@@ -29,15 +33,17 @@ def temporal_closeness_centrality(data):
 
     '''
 
-    pathdata = 0  # are shortest paths calculated
-    if isinstance(data, dict):
-        # This could be done better
-        if [k for k in list(data.keys()) if k == 'paths'] == ['paths']:
-            pathdata = 1
+    if tnet is not None and paths is not None: 
+        raise ValueError('Only network or path input allowed.')
+    if tnet is None and paths is None: 
+        raise ValueError('No input.')
     # if shortest paths are not calculated, calculate them
-    if pathdata == 0:
-        data = shortest_temporal_path(data)
+    if tnet is not None:
+        paths = shortest_temporal_path(tnet)
 
-    closeness = np.nansum(1 / np.nanmean(data['paths'], axis=2), axis=1) / (data['paths'].shape[1] - 1)
+    pathmat = np.zeros([paths[['from','to']].max().max()+1, paths[['from','to']].max().max()+1, paths[['t_start']].max().max()+1]) * np.nan     
+    pathmat[paths['from'].values,paths['to'].values,paths['t_start'].values] = paths['temporal-distance']
+
+    closeness = np.nansum(1 / np.nanmean(pathmat, axis=2), axis=1) / (pathmat.shape[1] - 1)
 
     return closeness