+++ teneto/networkmeasures/temporal_efficiency.py
@@ -5,15 +5,21 @@ Networkmeasures: Temporal Efficiency
 import numpy as np
 from .shortest_temporal_path import shortest_temporal_path 
 
-def temporal_efficiency(data, calc='global'):
+def temporal_efficiency(tnet=None, paths=None, calc='global'):
     r"""
     Returns temporal efficiency estimate. BU networks only.
 
     Parameters
     ----------
-    data: dict or array
-        If array, graphlet.
-        If dict, either contact representation or paths (output from shortest_temporal_path).
+    Input should be *either* tnet or paths. 
+
+    data : array or dict 
+    
+        Temporal network input (graphlet or contact). nettype: 'bu', 'bd'. 
+        
+    paths : pandas dataframe
+    
+        Output of TenetoBIDS.networkmeasure.shortest_temporal_paths
 
     calc : str
         Options: 'global' (default) - measure averages over time and nodes;
@@ -29,21 +35,23 @@ def temporal_efficiency(data, calc='global'):
 
     """
 
-    paths = 0  # are shortest paths calculated
-    if isinstance(data, dict):
-        # This could be calcne better
-        if [k for k in list(data.keys()) if k == 'paths'] == ['paths']:
-            paths = 1
+    if tnet is not None and paths is not None: 
+        raise ValueError('Only network or path input allowed.')
+    if tnet is None and paths is None: 
+        raise ValueError('No input.')
     # if shortest paths are not calculated, calculate them
-    if paths == 0:
-        data = shortest_temporal_path(data)
+    if tnet is not None:
+        paths = shortest_temporal_path(tnet)
+
+    pathmat = np.zeros([paths[['from','to']].max().max()+1, paths[['from','to']].max().max()+1, paths[['t_start']].max().max()+1]) * np.nan     
+    pathmat[paths['from'].values,paths['to'].values,paths['t_start'].values] = paths['temporal-distance']
 
     # Calculate efficiency which is 1 over the mean path.
     if calc == 'global':
-        eff = 1 / np.nanmean(data['paths'])
+        eff = 1 / np.nanmean(pathmat)
     elif calc == 'node' or calc == 'node_from':
-        eff = 1 / np.nanmean(np.nanmean(data['paths'], axis=2), axis=1)
+        eff = 1 / np.nanmean(np.nanmean(pathmat, axis=2), axis=1)
     elif calc == 'node_to':
-        eff = 1 / np.nanmean(np.nanmean(data['paths'], axis=2), axis=0)
+        eff = 1 / np.nanmean(np.nanmean(pathmat, axis=2), axis=0)
 
     return eff