+++ test/networkmeasures/test_shortesttemporalpath.py
@@ -3,54 +3,91 @@ import teneto
 import numpy as np
 import pytest 
 
-def test_sp_error():
-    G = np.zeros([3, 3, 4])
-    G[0, 1, [0, 2, 3]] = 1
-    G[0, 2, 1] = 1
-    G[1, 2, 3] = 0.5
-    G += G.transpose([1, 0, 2])
-    G = teneto.utils.set_diagonal(G, 1)
-    with pytest.raises(ValueError):
-        sp = teneto.networkmeasures.shortest_temporal_path(G, quiet=0)
+# def test_sp_error():
+#     G = np.zeros([3, 3, 4])
+#     G[0, 1, [0, 2, 3]] = 1
+#     G[0, 2, 1] = 1
+#     G[1, 2, 3] = 0.5
+#     G += G.transpose([1, 0, 2])
+#     G = teneto.utils.set_diagonal(G, 1)
+#     with pytest.raises(ValueError):
+#         sp = teneto.networkmeasures.shortest_temporal_path(G, quiet=0)
 
-def test_networkmeasures_stp():
-    # Make simple network
-    G = np.zeros([3, 3, 4])
-    G[0, 1, [0, 2, 3]] = 1
-    G[0, 2, 1] = 1
-    G[1, 2, 3] = 1
-    G += G.transpose([1, 0, 2])
-    G = teneto.utils.set_diagonal(G, 1)
-    sp = teneto.networkmeasures.shortest_temporal_path(G, quiet=0)
-    sp['paths'] = teneto.utils.set_diagonal(sp['paths'], 0)
-    paths_true = np.zeros(sp['paths'].shape)
-    # reminder dimord is from,to
-    paths_true[0, 1, 0] = 1
-    paths_true[0, 2, 0] = 2
-    paths_true[1, 0, 0] = 1
-    paths_true[1, 2, 0] = 2
-    paths_true[2, 1, 0] = 3
-    paths_true[2, 0, 0] = 2
-    paths_true[0, 1, 1] = 2
-    paths_true[0, 2, 1] = 1
-    paths_true[1, 0, 1] = 2
-    paths_true[1, 2, 1] = 3
-    paths_true[2, 1, 1] = 2
-    paths_true[2, 0, 1] = 1
-    paths_true[0, 1, 2] = 1
-    paths_true[0, 2, 2] = 2
-    paths_true[1, 0, 2] = 1
-    paths_true[1, 2, 2] = 2
-    paths_true[2, 1, 2] = 2
-    paths_true[2, 0, 2] = 2
-    paths_true[0, 1, 3] = 1
-    paths_true[0, 2, 3] = 1
-    paths_true[1, 0, 3] = 1
-    paths_true[1, 2, 3] = 1
-    paths_true[2, 1, 3] = 1
-    paths_true[2, 0, 3] = 1
-    assert (sp['paths'] == paths_true).all()
+# def test_networkmeasures_stp():
+#     # Make simple network
+#     G = np.zeros([3, 3, 4])
+#     G[0, 1, [0, 2, 3]] = 1
+#     G[0, 2, 1] = 1
+#     G[1, 2, 3] = 1
+#     G += G.transpose([1, 0, 2])
+#     G = teneto.utils.set_diagonal(G, 1)
+#     sp = teneto.networkmeasures.shortest_temporal_path(G, quiet=0)
+#     sp['paths'] = teneto.utils.set_diagonal(sp['paths'], 0)
+#     paths_true = np.zeros(sp['paths'].shape)
+#     # reminder dimord is from,to
+#     paths_true[0, 1, 0] = 1
+#     paths_true[0, 2, 0] = 2
+#     paths_true[1, 0, 0] = 1
+#     paths_true[1, 2, 0] = 2
+#     paths_true[2, 1, 0] = 3
+#     paths_true[2, 0, 0] = 2
+#     paths_true[0, 1, 1] = 2
+#     paths_true[0, 2, 1] = 1
+#     paths_true[1, 0, 1] = 2
+#     paths_true[1, 2, 1] = 3
+#     paths_true[2, 1, 1] = 2
+#     paths_true[2, 0, 1] = 1
+#     paths_true[0, 1, 2] = 1
+#     paths_true[0, 2, 2] = 2
+#     paths_true[1, 0, 2] = 1
+#     paths_true[1, 2, 2] = 2
+#     paths_true[2, 1, 2] = 2
+#     paths_true[2, 0, 2] = 2
+#     paths_true[0, 1, 3] = 1
+#     paths_true[0, 2, 3] = 1
+#     paths_true[1, 0, 3] = 1
+#     paths_true[1, 2, 3] = 1
+#     paths_true[2, 1, 3] = 1
+#     paths_true[2, 0, 3] = 1
+#     assert (sp['paths'] == paths_true).all()
 
+def test_sp_new(): 
+    G = np.zeros([5, 5, 10])
+    G[0, 1, [0, 2, 3, 8]] = 1
+    G[0, 2, [7, 9]] = 1
+    G[1, 2, [1]] = 1
+    G[1, 3, [5, 8]] = 1
+    G[2, 4, [2, 4, 9]] = 1
+    G[2, 3, [1, 3, 4, 7]] = 1
+    # CHeck output is correct length
+    paths_1step = teneto.networkmeasures.shortest_temporal_path(G,1)
+    paths_all = teneto.networkmeasures.shortest_temporal_path(G,'all')
+    dflen = np.prod(G.shape)-(G.shape[0]*G.shape[2])
+    assert len(paths_all) == dflen == len(paths_1step)
+    # Make sure these two are different
+    assert (paths_1step==paths_all).all()['temporal-distance'] == False
+    # Check first edge and make sure all works
+    paths_1step = teneto.networkmeasures.shortest_temporal_path(G,1,i=0,j=3,it=0)
+    paths_all = teneto.networkmeasures.shortest_temporal_path(G,'all',i=0,j=3,it=0)
+    paths_2step = teneto.networkmeasures.shortest_temporal_path(G, 2,i=0,j=3,it=0)
+    # Currently broken
+    # Appears something is going wrong with the 2nd step (I think) - wrong path include
+    # But also the temporal-distance measure is broken as it returns the max.
+    paths_1step_topo = teneto.networkmeasures.shortest_temporal_path(G, 2,i=0,j=3,it=0,minimise='topology')
+    assert (paths_all == paths_2step).all().all()
+    assert paths_all['topological-distance'].values == 3
+    assert paths_all['temporal-distance'].values == 2
+    assert paths_1step['topological-distance'].values == 3
+    assert paths_1step['temporal-distance'].values == 4
+    # Check a second edge
+    paths_all = teneto.networkmeasures.shortest_temporal_path(G,'all',i=0,j=3,it=3)
+    paths_1step = teneto.networkmeasures.shortest_temporal_path(G,'all',i=0,j=3,it=3)
+    assert paths_all['topological-distance'].values == 2
+    assert paths_all['temporal-distance'].values == 3
+    assert paths_1step['topological-distance'].values == 2
+    assert paths_1step['temporal-distance'].values == 3
+    # Check path is correct
+    assert paths_all['path includes'].values[0] == [[0, 1], [1, 3]]
 
 def test_networkmeasures_teff():
     # Test temporal efficiency
@@ -62,17 +99,21 @@ def test_networkmeasures_teff():
     G = teneto.utils.set_diagonal(G, 1)
     E = teneto.networkmeasures.temporal_efficiency(G)
     sp = teneto.networkmeasures.shortest_temporal_path(G)
-    E2 = teneto.networkmeasures.temporal_efficiency(sp)
+    E2 = teneto.networkmeasures.temporal_efficiency(paths=sp)
     assert E == E2
     # Matrix symmetric so nodal measure is same regardless of how you calculate paths
-    EN1 = teneto.networkmeasures.temporal_efficiency(sp, calc='node_to')
-    EN2 = teneto.networkmeasures.temporal_efficiency(sp, calc='node_from')
+    EN1 = teneto.networkmeasures.temporal_efficiency(paths=sp, calc='node_to')
+    EN2 = teneto.networkmeasures.temporal_efficiency(paths=sp, calc='node_from')
     assert all(EN1 == EN2)
     # Change G so matrix is directed now index 0 should be less efficient in "from" (this feature isn't implemented in teneto yet)
     #G[0,2,1] = 0
     #EN1 = teneto.networkmeasures.temporal_efficiency(G,calc='node_to')
     #EN2 = teneto.networkmeasures.temporal_efficiency(G,calc='node_from')
-
+    # bad inputs that will raise errors
+    with pytest.raises(ValueError):
+        teneto.networkmeasures.temporal_efficiency(G, paths=sp)
+    with pytest.raises(ValueError):
+        teneto.networkmeasures.temporal_efficiency()
 
 def test_reachrat():
     # Test temporal efficiency
@@ -84,13 +125,34 @@ def test_reachrat():
     G = teneto.utils.set_diagonal(G, 1)
     R = teneto.networkmeasures.reachability_latency(G)
     sp = teneto.networkmeasures.shortest_temporal_path(G)
-    R2 = teneto.networkmeasures.reachability_latency(sp)
+    R2 = teneto.networkmeasures.reachability_latency(paths=sp)
     assert R == R2
     # Matrix symmetric so nodal measure is same regardless of how you calculate paths
     RN1 = teneto.networkmeasures.reachability_latency(G, calc='nodes')
-    RN2 = teneto.networkmeasures.reachability_latency(sp, calc='nodes')
+    RN2 = teneto.networkmeasures.reachability_latency(paths=sp, calc='nodes')
     assert np.all(RN1 == RN2)
-    paths = teneto.utils.set_diagonal(sp['paths'], 0)
-    # Rglobal is average of the longest shortest path.
-    RS = paths.max(axis=0).mean(axis=1).mean()
-    assert RS == R
+    # bad inputs that will raise errors
+    with pytest.raises(ValueError):
+        teneto.networkmeasures.reachability_latency(G, paths=sp)
+    with pytest.raises(ValueError):
+        teneto.networkmeasures.reachability_latency()
+        
+
+def test_bet(): 
+    G = np.zeros([5, 5, 3])
+    G[0, 1, [0, 2,]] = 1
+    G[1, 2, [1]] = 1
+    G[2, 4, [2]] = 1
+    G[2, 3, [1]] = 1
+    # CHeck output is correct length
+    bet_time = teneto.networkmeasures.temporal_betweenness_centrality(G)
+    bet_global = teneto.networkmeasures.temporal_betweenness_centrality(G,calc='global')
+    assert (np.mean(bet_time,axis=1) == bet_global).all()
+    sp = teneto.networkmeasures.shortest_temporal_path(G)
+    with pytest.raises(ValueError):
+        teneto.networkmeasures.reachability_latency(G, paths=sp)
+    with pytest.raises(ValueError):
+        teneto.networkmeasures.reachability_latency()
+    #calculated by looking at G
+    assert bet_global[1] == (3*1/((G.shape[0]-1)*(G.shape[0]-2)))/G.shape[-1]
+    assert bet_global[2] == (6*1/((G.shape[0]-1)*(G.shape[0]-2)))/G.shape[-1]